requirements.txt
strands-agents
strands-agents[ollama]
boto3
python-dotenv
